:bug: Fix: add card component into form and title outside like Design
